requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
drf-yasg==1.21.8
environs==11.0.0
flake8==7.1.2
flake8-docstrings==1.7.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
marshmallow==3.23.1
mccabe==0.7.0
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
setuptools==75.1.0
snowballstemmer==2.2.0
sqlparse==0.5.1
typing==3.7.4.3
tzdata==2024.2
uritemplate==4.1.1
